+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This repo consists of two packages - the React client and the Python server. A change to either repo should trigger an update in the versions for both to ensure a consistent changelog in this repo.
 
+## [1.10.3] - 2022-01-28
+
+### Fixed
+
+* A bug which caused boolean parameters to display "false -> false" when unchecked.
+* The SMF tax-based payment now is paid to basic/intermediate/starter rate payers, rather than just not higher and additional rate payers.
+
 ## [1.10.2] - 2022-01-26
 
 ### Changed